move RulesExtra (only struct, keep behavior)
